Separate indexing module
Separated indexing.rs into square.rs, pieces.rs and castling_rights.rs.
Also removed square.rs and pieces.rs from public API.
