+++ position/Cargo.toml
@@ -5,5 +5,5 @@ edition = "2021"
 
 [dependencies]
 bitflags = "2.3.1"
-const-random = "0.1.18"
 strum = { version = "0.27.1", features = ["derive"] }
+const-random = "0.1.18"
