+++ position/src/prelude.rs
@@ -1,6 +1,3 @@
-pub use crate::indexing::{
-    CastlingRights, Color, File, NegDiag, Piece, PosDiag, Promotion, Rank, Square,
-};
-pub use crate::move_generation::{ChessMove, ChessMoveHint, ChessUnmove, LanMove};
-pub use crate::position::{ParseFenError, Position};
+pub use crate::move_generation::{ChessMove, ChessMoveHint, ChessUnmove, LanMove, PackedChessMove};
+pub use crate::position::{CastlingRights, ParseFenError, Position};
 pub use crate::position_hash::PositionHash;
